Update Main.java
This is a Java program that implements a class schedule system. It allows users to add, remove, and display classes, as well as sort the schedule based on class times. The program uses object-oriented programming concepts and utilizes classes such as Schedule, Class, Subject, and Teacher. User input is obtained through the Scanner class.